script/requirements.txt
beautifulsoup4==4.12.3
GitPython==3.1.43
lxml==5.2.2
matplotlib==3.9.0
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
PyYAML==6.0.1
Requests==2.32.2
selenium==4.21.0
xlrd==2.0.1